feature: exFAT working
